Switch the node uom based on the custom parameter 'Units'.
Convert the values and change the uom if the Units is switch to
imperial (or really, non-metric).

However, this does not change the actual node definition used so
this doesn't really work and shouldn't be used at this point. It
appears that Polyglot doesn't support the ability to change a node's
node definition once it's been created.