Clear out old rain values on start.
Compare the saved time with the current time for each rain accumulator
and clear those that have expired.

Use week number instead of day of week to track weekly rain.

Hourly rain was not resetting to zero for the next hour.

Also, fix pressure trend values and don't try to convert them when US
units are selected.