Experimental code to query WF servers for meta data
Query the WF servers for the station elevation, the air sensor height
above ground and the user's preferred units.  If found, use these
values.

Note that the units are used to determin the same high level units
that the node server configuration allows (us, uk, metric) and not
the individual sensor units.